+++ src/gui/qt/plugin_editor/models/location_table_model.cpp
@@ -59,7 +59,7 @@ QVariant LocationTableModel::data(const QModelIndex& index, int role) const {
   }
 
   const auto& element =
-      index.row() < nonUserMetadata.size()
+      index.row() < static_cast<int>(nonUserMetadata.size())
           ? nonUserMetadata.at(index.row())
           : userMetadata.at(index.row() - nonUserMetadata.size());
 
@@ -99,7 +99,7 @@ Qt::ItemFlags LocationTableModel::flags(const QModelIndex& index) const {
     return Qt::ItemIsEnabled;
   }
 
-  if (index.row() < nonUserMetadata.size()) {
+  if (index.row() < static_cast<int>(nonUserMetadata.size())) {
     return QAbstractItemModel::flags(index);
   }
 
@@ -113,7 +113,7 @@ bool LocationTableModel::setData(const QModelIndex& index,
     return false;
   }
 
-  if (index.row() < nonUserMetadata.size() ||
+  if (index.row() < static_cast<int>(nonUserMetadata.size()) ||
       index.column() > columnCount() - 1) {
     return false;
   }
@@ -133,7 +133,7 @@ bool LocationTableModel::setData(const QModelIndex& index,
 bool LocationTableModel::insertRows(int row,
                                     int count,
                                     const QModelIndex& parent) {
-  if (row < nonUserMetadata.size() || row > rowCount()) {
+  if (row < static_cast<int>(nonUserMetadata.size()) || row > rowCount()) {
     return false;
   }
 
@@ -151,7 +151,7 @@ bool LocationTableModel::insertRows(int row,
 bool LocationTableModel::removeRows(int row,
                                     int count,
                                     const QModelIndex& parent) {
-  if (row < nonUserMetadata.size() || row > rowCount() ||
+  if (row < static_cast<int>(nonUserMetadata.size()) || row > rowCount() ||
       row + count > rowCount()) {
     return false;
   }
